Start working on filesystem
