+++ packages/wasi-motor/lib/snapshot-preview1.ts
